backend/requirements.txt
requests
beautifulsoup4
python-dotenv
pymongo
fastapi[all]